pyproject.toml
[project]
name = "cracked-leetcode-junkie-bot"
version = "0.1.0"
description = "A Discord bot with AI capabilities and message indexing for context-aware responses"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.13.2",
    "aiosqlite>=0.22.0",
    "discord-py>=2.6.4",
    "google-genai>=1.55.0",
    "python-dotenv>=1.2.1",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.14.9",
]

[tool.ruff]
indent-width = 2

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "B"]
ignore = ["B011", "E402", "E741"]
